+++ src/main/java/com/test/models/PurchaseRequisitionItems.java
@@ -37,7 +37,7 @@ public class PurchaseRequisitionItems {
 	private Double quantity;
 
 	@Column
-	private Integer unit_price;
+	private Double unit_price;
 
 	@Column
 	private Double total_price;
@@ -132,11 +132,11 @@ public class PurchaseRequisitionItems {
 		this.quantity = quantity;
 	}
 
-	public Integer getUnit_price() {
+	public Double getUnit_price() {
 		return unit_price;
 	}
 
-	public void setUnit_price(Integer unit_price) {
+	public void setUnit_price(Double unit_price) {
 		this.unit_price = unit_price;
 	}
