+++ src/main/java/com/test/models/StockDeliveryItems.java
@@ -35,7 +35,7 @@ public class StockDeliveryItems {
 	private Double quantity;
 
 	@Column
-	private Integer unit_price;
+	private Double unit_price;
 
 	@Column
 	private Double total_price;
@@ -112,11 +112,11 @@ public class StockDeliveryItems {
 		this.quantity = quantity;
 	}
 
-	public Integer getUnit_price() {
+	public Double getUnit_price() {
 		return unit_price;
 	}
 
-	public void setUnit_price(Integer unit_price) {
+	public void setUnit_price(Double unit_price) {
 		this.unit_price = unit_price;
 	}
